Update agency.css
Hi, I'm assuming the nav bar is part of the home page and overall design of the website. Tracy and I have been discussing what we want to do with it and we have some ideas already so if you guys could try not to touch the code for the index or the CSS that'd be great. Thanks! :)